send messages at 15Hz

+++ GamesProgramming/Ex_Audio/GameState.h
@@ -20,7 +20,7 @@ public:
 	~GameState();
 
 protected:
-	float frame;
+	unsigned short frame;
 
 	AudioManager audioMgr;
